refactor(user-dialog): moving toward more general and reusable header and footer modal component
