+++ app/dashboard/src/components/Dialog/Icons.tsx
@@ -1,0 +1,21 @@
+
+
+import { chakra } from '@chakra-ui/react';
+import {
+  PencilIcon,
+  UserPlusIcon,
+} from '@heroicons/react/24/outline';
+
+export const AddIcon = chakra(UserPlusIcon, {
+  baseStyle: {
+    w: 5,
+    h: 5,
+  },
+});
+
+export const EditIcon = chakra(PencilIcon, {
+  baseStyle: {
+    w: 5,
+    h: 5,
+  },
+});
