+++ app/dashboard/src/components/Dialog/ModalHeader.tsx
@@ -3,11 +3,7 @@ import { Text, HStack, ModalHeader } from '@chakra-ui/react';
 import { ChakraIcon } from 'types/ChakraIcon';
 import { Icon } from '../Icon';
 
-type DialogModalHeaderProps = {
-    HeaderIcon: ChakraIcon, title: string
-}
-
-export const DialogModalHeader = ({ HeaderIcon, title }: DialogModalHeaderProps) => {
+export const UserDialogModalHeader = ({ HeaderIcon, title }: { HeaderIcon: ChakraIcon, title: string }) => {
   return (
     <ModalHeader pt={6}>
       <HStack gap={2}>
